integrate mmseqs2 pipeline

+++ alphafold/data/tools/utils.py
@@ -22,14 +22,14 @@ from absl import logging
 
 
 @contextlib.contextmanager
-def tmpdir_manager(base_dir: Optional[str] = None):
+def tmpdir_manager(base_dir: Optional[str] = None, prefix: Optional[str] = None, delete: Optional[bool] = True):
   """Context manager that deletes a temporary directory on exit."""
-  tmpdir = tempfile.mkdtemp(dir=base_dir)
+  tmpdir = tempfile.mkdtemp(dir=base_dir, prefix=prefix)
   try:
     yield tmpdir
   finally:
-    shutil.rmtree(tmpdir, ignore_errors=True)
-
+    if delete:
+      shutil.rmtree(tmpdir, ignore_errors=True)
 
 @contextlib.contextmanager
 def timing(msg: str):
